refactor: improve types directory naming
